Polishes authentication support
- Libvirt drivers release memory allocated
for username and password strings. We are using strdup to duplicate
the strings and allow Libvirt drivers to free the  memory.

- segfault-handler module was added to aid development by
printing more detailed information when segfaults occur.

- Removes use of C++ standard library to keep the node-libvirt code
consistent with Node.js core code.

- Uses Hypervisor own instance to temporary store username and password
as opposed to creating a new class for those properties only.

- Makes authentication callback a private member of the Hypervisor class
to keep consistency with the other callback functions present in this class.

Authencation was tested against VMware VSphere 5.5 but commented in the tests
as Libvirt Test driver does not support this function.