refactor(destroy): don't clear pointers in desetroy scenarios
We used to call Clear() on the LibVirtHandle in each of these
cases, but the documentation explicitly states that the pointers
are _not_ freed when these methods are called. We should stick as
closely as possible to the libvirt API, and only clean these up
when a domain or hypervisor is explicitly deleted.