refactor(domain): reimplement RegisterEvent/UnregisterEvent
These methods are now implemented using RunAsync, as well as some
cleanups to make the callbacks in line with the rest of the code
base's style. RegisterEvent now defaults to taking an event as its
first parameter, while still supporting the old form of passing in
an object with an `evtype` property.