fix(error): add proper context to error messages
NLVAsyncWorkerBase was previously modified to include context for
error location when libvirt operations failed. This didn't
translate over to the most recent async worker work, and therefore
error context was consistently showing as being called from
`nlv_async_worker.cc:58`, or the callsite of the error callback of
the async worker. This patch corrects this behavior