refactor(promisificaton): support nodeback + promise first class
Originally all methods exported by the bindings supported
nodebacks, and we wrapped all of those with bluebird to provide
promisified versions of each method with an `Async` suffix. This
patch modifies that approach to support calling every method with
the same name, and determining if a Promise should be returned by
whether the user has provided a nodeback as the final argument to
the method.